training/requirements.txt
tensorflow>=2.16.0
numpy>=1.24.0
pandas>=2.0.0
numba>=0.59.0
python-chess>=1.999
onnx>=1.14.0
onnxruntime>=1.15.0
tf2onnx>=1.16.0
matplotlib>=3.7.0
